requirements.txt
# Core dependencies for StyleTTS2
torch>=1.8.0
torchaudio>=0.8.0
numpy
librosa
soundfile
munch
pyyaml
click

# FastAPI dependencies  
fastapi
uvicorn
python-multipart

# Audio processing
phonemizer

# Additional dependencies
tensorboard